[feat] implement getSubscribedSW feature
- seperate Table from tabs to use Suspense
- In table, implement getSubscribedSW feature
